+++ github_mcp_server/src/ambient/agent_injector.py
@@ -14,7 +14,7 @@ import json
 
 # Импортируем из родительского пакета
 sys.path.append(str(Path(__file__).parent.parent))
-from ..core.base import BaseWizard
+from ..core.base_wizard import BaseWizard
 
 class AgentInjector(BaseWizard):
     """Инжектор промптов в cursor-agent"""
