+++ github_mcp_server/src/ambient/ambient_agent.py
@@ -12,8 +12,8 @@ from typing import Optional
 
 # Импортируем из родительского пакета
 sys.path.append(str(Path(__file__).parent.parent))
-from ..core.base import BaseWizard
-from ..core.env import EnvManager
+from ..core.base_wizard import BaseWizard
+from ..core.env_manager import EnvManager
 
 # Импортируем ambient компоненты
 from .event_system import EventSystem, Event, EventType
