+++ github_mcp_server/src/ambient/event_handlers.py
@@ -10,7 +10,7 @@ from typing import TYPE_CHECKING
 
 # Импортируем из родительского пакета
 sys.path.append(str(Path(__file__).parent.parent))
-from ..core.base import BaseWizard
+from ..core.base_wizard import BaseWizard
 from .event_system import Event, EventType
 
 # Избегаем циклических импортов
