+++ github_mcp_server/src/ambient/event_system.py
@@ -15,7 +15,7 @@ from pathlib import Path
 
 # Импортируем из родительского пакета
 sys.path.append(str(Path(__file__).parent.parent))
-from ..core.base import BaseWizard
+from ..core.base_wizard import BaseWizard
 
 class EventType(Enum):
     """Типы событий в системе"""
