+++ github_mcp_server/src/core/base_wizard.py
