+++ github_mcp_server/tests/preflight/check_github.py
@@ -1,4 +1,4 @@
-from src.core.env import EnvManager
+from src.core.env_manager import EnvManager
 from src.setup.github_setup import GitHubSetup
